REFACTOR `MockRedisCacheTest` into traits for each area of functionality
ADD setting the 'database.redis.client' config value in `MockRedisCacheTest` so that multiple client config values can be used.
