+++ tests/Unit/Traits/MissingTest.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Sfneal\Helpers\Redis\Tests\Unit\Traits;
+
+use Sfneal\Helpers\Redis\RedisCache;
+
+trait MissingTest
+{
+    public function test_missing()
+    {
+        $key = 'bos-99';
+        $expected = RedisCache::missing($key);
+
+        $this->assertTrue($expected == true);
+    }
+}
